Previous fix wasn't so great, abandon the CloudCityHandler
Makes no sense for the LoggingService to just have a handler which it'll pass over to our methods.
Since our methods instantiate a new internal handler for the update thread, might as well also
instantiate a new HandlerThread which will force the updates to a background thread (away from the main thread).

So this fix is more resiliant to problems than the previous one (which also didn't work since the handler was never instantiated)