Enable logging and CC logging on app start
Well, the MainActivity wasn't changed much - the MainActivityExtensions call
was moved downwards, after the DataProvider is initialized, and was also extended
with the DataProvider itself.

The MainActivityExtensions entry method itself sets a listener on the SharedPrefGrouper
for a particular key - and refreshes all of the stuff in the DataProvider when
it happens.

The second methods just directly writes 'true' for the two keys:
the general 'enable logging' and 'enable citycloud logging' keys,

Finally, the Constants themselves got two new keys that we're setting by directly writing
true to them.

EDIT: and after all that, remove the logging_sp listener we've initially set because we won't need it anymore.