Get rid of the potential discrepancy between SharedPref and out Repository
When the app starts, the SharedPref is prefilled from hardcoded parameters from the
Repository. The networking requests still use the value from SharedPref unless they're
empty; in case they're empty we use the ones from Repo. However, if the user goes into the
Settings and changes the server URL/token to something else, we will have a discrepancy
between the repository and what's stored in the SharedPref until the app is restarted
which is when the repo will sync to what's in the SharedPrefs.

Use the SharedPrefListener approach to listen to these important key changes, and immediatelly
update the repo if/when they change.