Change when we turn on the CloudCity logging
So, a short testing session shown the gullibility of the previous approach.
Namely, everything works great when you start granting permissions the moment
the permission dialogs show up.

But if you wait for a while, the server fails never stop.

So to avoid all that hassle, lets automatically turn on the CC logging only when
it's safe to do so, which is when we've received the ACCESS_FINE_LOCATION permission.

Now the app waits for the user's permission inputs *before* it automatically turns on
the CC logging and all server sends (apart from maybe the first one) immediatelly start
and succeed.