Get rid of the potential discrepancy between SharedPref and out Repo (#15)
When the app starts, the SharedPref is prefilled from hardcoded parameters from the Repository. The networking requests still use the value from SharedPref unless they're empty; in case they're empty we use the ones from Repo. However, if the user goes into the Settings and changes the server URL/token to something else, we will have a discrepancy between the repository and what's stored in the SharedPref until the app is restarted which is when the repo will sync to what's in the SharedPrefs.

Use the SharedPrefListener approach to listen to these important key changes, and immediatelly update the repo if/when they change.

EDIT: also ended up fixing a big issue with when the CloudCity logging is started, since it requires FINE_LOCATION permission in order to work fine.
Starting it before that permission is granted kinda gets it stuck into a bad state, so that (apart from other bugs) was fixed.


* Get rid of the potential discrepancy between SharedPref and out Repository

When the app starts, the SharedPref is prefilled from hardcoded parameters from the
Repository. The networking requests still use the value from SharedPref unless they're
empty; in case they're empty we use the ones from Repo. However, if the user goes into the
Settings and changes the server URL/token to something else, we will have a discrepancy
between the repository and what's stored in the SharedPref until the app is restarted
which is when the repo will sync to what's in the SharedPrefs.

Use the SharedPrefListener approach to listen to these important key changes, and immediatelly
update the repo if/when they change.

* Fix the isBlank() shared pref listener CR comment

* Change when we turn on the CloudCity logging

So, a short testing session shown the gullibility of the previous approach.
Namely, everything works great when you start granting permissions the moment
the permission dialogs show up.

But if you wait for a while, the server fails never stop.

So to avoid all that hassle, lets automatically turn on the CC logging only when
it's safe to do so, which is when we've received the ACCESS_FINE_LOCATION permission.

Now the app waits for the user's permission inputs *before* it automatically turns on
the CC logging and all server sends (apart from maybe the first one) immediatelly start
and succeed.