Database patches / updates
Nothing too significant, but the Results database was package-private
and we need to access it from our own cloudcity package.

The Iperf3RunResultDao was severely lacking one more method for fetching
the latest run result as live data (rather than getting it via UID)
so that method was also added in.