Everything works fine now
The Iperf3Monitor was finished and is working as expected, parsing the iperf3 results and posting them back
to the listener.

The Iperf3/Iperf3Parser had to be slightly extended, with the completion listener (in case you don't want
to listen to every single property change) as well as a static instantiation method.

And a new POJO dataholder was made to hold all of these upload/download metrics.