Fixed two outstanding flaws that didn't happen all the time...
The two flaws were a significant poke in the eye though.
1. the parsing thread would sometimes keep spinning forever ...
2. sometimes, or rather alot of the times the parser would argue
and throw an exception about parsing errors because the file deletion
that happens after the parsing thread was stopped would literally
pull the file from under the parser's feet.

And these two things had to be addressed so that we get a 10/10
"parsing stops when we really thing it stopped, and delete the file after that moment"
results, rather than the weird one-off or 3/10-off results affected by internal
race conditions and whatnot.

Now, the parser doesn't whine anymore. coderabbit might, but i'm not changing anything else.
this is done for good now.