Update LoggingServiceExtensions
Before, we used to pick up all information (and RSSNR) from the "registered cell" information.
Whatever that is, on phones with SIMs, including my wife's phone, I can clearly see that on the
Details page, the Cell Information says "N/A" for RSSNR while the "Signal Strength Information"
clearly says -1 for RSSNR.

So, to find a common ground between the two, we'll initialize our data model for sending from
the registered cell, and update it with signal information if possible.

This should net us the mix of both, with the least possible amount of nulls in between.