Adapting iperf3 to get data easily (#20)
* Add prefill button to the Iperf3Fragment

Will prefill with CloudCity Iperf3 server and a random port
in the allowed port range.

* Checkpoint commit - part1

Extracted the 'external' Iperf3 parsing to the CloudCityUtil class;
still need to add the whole thread and alot of other things
(as well as minimize the changes in other files that i had to modify since the build will most likely fail)

but this is a good checkpoint none the less.

* Checkpoint commit - part2

Decided that CloudCityUtil isn't the best place for this,
since this whole mechanism isn't just one method so it needs a class of it's own.

Migrated the code to the Iperf3Monitor, added threading, but results are still no good.

Still, I know what needs to be done now.

* Change the iperf3 prefill to fill in duration rather than interval

Whoops, my bad. Was thinking about duration and somehow put the wrong field in by mistake.

* Iperf3Monitor update

Seems to be parsing the whole file but... for some reason, the parsing thread never stops.
That needs to be looked into.

* Database patches / updates

Nothing too significant, but the Results database was package-private
and we need to access it from our own cloudcity package.

The Iperf3RunResultDao was severely lacking one more method for fetching
the latest run result as live data (rather than getting it via UID)
so that method was also added in.

* Everything works fine now

The Iperf3Monitor was finished and is working as expected, parsing the iperf3 results and posting them back
to the listener.

The Iperf3/Iperf3Parser had to be slightly extended, with the completion listener (in case you don't want
to listen to every single property change) as well as a static instantiation method.

And a new POJO dataholder was made to hold all of these upload/download metrics.

* Change logging levels, speed up parsing even more

* Addresing CR comments

* Fixed two outstanding flaws that didn't happen all the time...

The two flaws were a significant poke in the eye though.
1. the parsing thread would sometimes keep spinning forever ...
2. sometimes, or rather alot of the times the parser would argue
and throw an exception about parsing errors because the file deletion
that happens after the parsing thread was stopped would literally
pull the file from under the parser's feet.

And these two things had to be addressed so that we get a 10/10
"parsing stops when we really thing it stopped, and delete the file after that moment"
results, rather than the weird one-off or 3/10-off results affected by internal
race conditions and whatnot.

Now, the parser doesn't whine anymore. coderabbit might, but i'm not changing anything else.
this is done for good now.