Convert SharedPreferencesGrouper listener map to a multimap. Kind of.
Earlier, the SharedPreferencesGrouper 'spMap' - the map holding all of the
'SharedPreferences type' keys and their particular listeners - was a 1:1 plain
HashMap.

This approach utterly falls into pieces because even the original code sets a few
OnSharedPreferenceChangeListener to the same key, thus overwriting each other all the time.

While this may not still work so great, as we'd might have to implement a generic
'callback dispatcher' - a 'shared' listener that registers to all of these shared preferences
and simply fires the callback to all of the listeners in the map, it *does* seem to work
and does the job so far. I think.

Either way, this was one of the problems involved in why the server was never changing when it was changed
in the app while the app was running.