Fix retrofit
Ok, so while the server URL **wasn't** hardcoded in the Retrofit client,
the retrofit would get initialized only once (with the prefilled URL) and would
simply ignore all the changed URLs passed to it because it already had an instance.

Now, instead of just checking whether we already have an initialized client, we also check
whether the new URL differs from what the last initialized client was registered to (it's base URL)
and if they differ - we reinitialize a new Retrofit client.

This commit (change) lets us change servers on the fly while the app is running.