GPSMonitor patch - now monitors last speed and triggers iperf3 tests
Monitors last speed as measured by the GPS (LocationManager) every second,
if the speed has been under THRESHOLD_VALUE (5) for THRESHOLD_DURATION (5000ms)
it'll fire a callback and reset the count.

It's internal ValueMonitor will check the value every VALUE_MONITOR_INTERVAL (500ms)

Once the value has been under the desired value for at least the desired time,
the GPSMonitor was also set to initiate the new 'default' 15 sec test so we can check
how the whole mechanism functions and operates, and if it works.