Turns out this idea wasn't as great as I originally thought it'd be
Sure, we update location much more frequently than DataProvider and end up being
less coupled with their codebase but... our GPSMonitor also gets a 0.0 in the start.

So I'll just push all the changes *except* the GPSMonitor, since that change is good,
and we'll keep that one.

These ones... can sit and wait for a better decision later on after the demo.