Make flavors for different server environments (#30)
This PR introduces 3 new flavors, for 'staging', 'production' and 'demo' server. It slightly changes up where the prefill data is fetched from, and will most likely also have to fix the workflow.

But then we'll end up with 3 debug and 3 release builds, each pair targetting a different server environment (and still have it changable from settings tho)

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **New Features**
	- Introduced product flavors for different server environments: Staging, Demo, and Production.
	- Enhanced server URL and token management based on the selected build variant.

- **Bug Fixes**
	- Improved error handling for unknown build variants during initialization.

- **Improvements**
	- Updated output file naming convention to reflect the current build flavor.
	- Enhanced CI process for more flexible APK file handling.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->



* Make new 3 build flavors, one for each server environment

So now we have the 'demo', 'staging' and 'production' build flavors.

* Update CloudCityParamsRepository to include new server domains

With these new flavors in play, each defining a separate BuildConfig field,
we can now differentiate which server to actually use in the CloudCityParamsRepository;

So a new method to select the URL and corresponding (mine) token based on these booleans
has also been put in.

* Reorder flavors by changing their names

The new names are kinda sucky, but it's the only way to control their order in Android Studio

I really don't want 'demo' to be the default build variant, period. I want staging to be the default one.

* Make better names for built APKs

Give each badly-named flavor a nicer name to use in the APK filename.

* New (re)implementation of the spdxSbom gradle task

* Fix the AndroidCI workflow to adapt to new flavors

Now there's an extra flavorName folder in between the old
app\build\outputs\apk\debug
and the new
app\build\outputs\apk\a_staging\debug

so the workflow needs to look like this
app\build\outputs\apk\**\debug

to be compatible with any other flavors we might end up putting in

* Add a TODO to delete these tokens when we have a login system