Add the new sensitive permission in the manifest
Only system-level apps can use this, so...

Either way, the only way to get it is to whitelist the app like so:

The app must be allowlisted in the privapp-permission.xml file, which is located in the /system/etc/permissions/ directory on the device.

This file is typically managed by the device manufacturer or carrier, and you would need their cooperation to modify it.

So that's that. It's there, and we *should* be able to obtain it once whitelisted since
we'll be using a custom version of AOSP (hence we'll be vendors so to say)