Fix GPSMonitor initialization
Well, all is well and great on the initial launch; the user is prompted for permissions,
we react to them being granted/rejected and everything runs great.

However on subsequent app runs, we don't go trough that same code path, and we ended up
not initializing some cruicial components for sending CloudCity data, namely the GPSMonitor

So, added checking whether we have teh permission and initializing it separately from the
"permissions received" callback, and the code now properly initializes it if we had the permission
without having to request it first.