Implemented test run throttling in Iperf3Monitor
Since the iperf3 tests eat away too much data, we need to throttle them;
but since it's not the API requests that are eating data but the iPerf3 tests
themselves - we need to throttle how often do we run test.

The initial implementation of throttling has been put in into the Iperf3Monitor.

The remaining part of work is to add customizability of this throttling time, in seconds
(or minutes) to the settings screen.