Really implement throttling in Iperf3Monitor
The initial idea was good, but lacked a few changes here and there.
Now, the THROTTLING_THRESHOLD_IN_SECONDS parameter is configurable
and really determines whether the iperf3 run request will be discarded
or actually ran and uploaded to the server.

It's also parametrizable through the settings screen where all other
cloud city things are put in, but those will be commited later.