Null-guarded DataProvider in the model that iperf3 sending uses
The method getRegisteredCellInformationUpdatedBySignalStrengthInformation()
has actually run on a NPE with the DataProvider. Due to weird issues/circumstances,
sometimes it's just null for whatever OMNT inexplicable reason.

If that happens, we will just send the iperf3 data without the measurements, but
as long as DataProvider is non-null, we'll probably get some cell info and attach them
the the iPerf3 results to be sent to the API.