Put in a workaround so it can still work without signal strength
Well, my test device doesn't have a SIM, therefore the signal strength is null,
which always caused NPEs whenever anything was tried to be queried.

So in that case, if signalStrength is null, i've put in a hack that makes it use
currentCell for both of the values (or rather, so that the currentCell is treated
as both the currently-registered cell and signal strength) so that sending to the server
functionality is retained.