Change iperf3 test to last 10secs (#48)
Issue link:
https://github.com/Cloud-City-RS/cloud_city_platform/issues/760

Since our old default automated iperf3 test duration was 15, and we had to lower it to 10, this PR adjusts the few necessary constants to make that happen. Also, the old hardcoded duration of `"15"` was extracted into a new CloudCityConstants constant value.

On top of that, since we also had to change the speed threshold, those two constants were moved out of the `GPSMonitor` into the constants file, renamed for more descriptive naming to fit with the rest of the "automated iperf3 test"-related constants, and were put to use.

Finally, the iperf3 constants got their clarifying javadoc to distinguish between values used for prefilling the iperf3 screen and what's used for automated tests when the movement speed is below threshold.

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **New Features**
	- Introduced new constants for iPerf3 testing configurations, enhancing settings and automated testing capabilities.
	- Updated the `Iperf3Monitor` to initiate dynamic automated tests based on new constants.

- **Bug Fixes**
	- Improved threshold handling in `GPSMonitor` to utilize updated constants for more accurate monitoring.

- **Documentation**
	- Updated documentation throughout the codebase to reflect new constants and method changes for clarity.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->


* Reduce default iperf3 automated test duration to 10 from 15

Added a new constant to declare the duration we should use in our CloudCityConstants file,
rather than using the hardcoded "15" as before.

* Extract constants for starting test out of GPSMonitor into CloudCityConstants

Extracted the 'value threshold' and 'duration threshold' thresholds into the CloudCityConstants,
renamed them to be more appropriate and descriptive, added javadoc, and changed the speed threshold
value to be 1.8 to better align with 5km/h requirement (rather than the previous 5m/s which was 18km/h)

* Address own CR comment about javadocs

* Adjust the speed thresshold to be lower than what i initially thought

Since 1m/s is 3.6km/h, and 5km/h is ~1.388m/s (and not 1.8), lets adjust it to use
1.4m/s instead, rather than 6.5km/h what the old one was representing.

Somehow i misread 5/3.6 to be 1.8 instead of 1.38, my bad.