Fix EARFCN parsing
Turns out, the "bands" are obtained as a String. And this string isn't just a number,
but rather something in the form of "[]" or "[3]" or "[20]"

So, we need to sanitize the bandsString of all brackets, and *then* parse that integer
before setting it and sending it to the serverside.

This commit implements that sanitizing and the requests are going through nicely, with all
the things that the 1-second spammer puts in the "cell_info" parameter.