Logging indirection (#49)
Issue link:
https://github.com/Cloud-City-RS/cloud_city_platform/issues/781

The sole point of this PR is to do two things:
1) replace default android's `Log.` calls with our own indirection class which can forward it to Sentry, Timber and other loggers we may end up installing
2) to verify and ensure that logs do get sent to sentry, and ensure that we use manually typed-out logtags rather than the `getClass().getSimpleName()` which will break the moment we obfuscate the codebase.

oh and also to get rid of `Log.x()` calls and replace them with `CloudCityLogger.x()` calls instead.

Whether I should have ended up using CloudCityLogger in the OMNT's part of the codebase is debatable; perhaps those could have been `OMNTLogger` which would apply stricter logging rules. But, lets see how much crap it ends up spewing on sentry and whether we can even use the logs we end up getting there before reducing the bulk.

Oh right, this PR also implemented timber and connected it with sentry :D or should have connected it to sentry at least.
EDIT: since timber+sentry = useless, we got rid of that and are waiting to implement a backend call to send these logs to


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit


- **New Features**
	- Introduced a new logging utility, `CloudCityLogger`, enhancing the application's logging capabilities across various components.

- **Bug Fixes**
	- Improved error logging by incorporating exception details in log statements throughout the application.

- **Chores**
	- Updated logging statements in multiple classes to use `CloudCityLogger` instead of the standard Android `Log` class, ensuring consistency in the logging framework.
	- Adjusted import statements to remove `android.util.Log` and include `cloudcity.util.CloudCityLogger` where necessary.

- **Documentation**
	- Added import statements for `CloudCityLogger` in relevant files to support the new logging mechanism.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->


* Timber initialization helper class in kotlin

Since timber 5.0.1 seems to be kotlin-only and it has a known issue with java-only projects
I tried initializing it with a kotlin helper class, which seemed to work fine, but ran into
other types of problems...

So rather than fucking with that by either trying to downgrade kotlin to 1.6 or whatnot, lets just
downgrade Timber to 4.7.1 which seems to work with Java, and then see if it works with SentryTimberTree...

This commit just saves this file before i nuke it out of my local repo.

* Nuke the ktx package out of existence

* Timber initialization and integration with Sentry

At least I hope this is correctly integrated now. We'll see if that's enough
when I start replacing all logs and causing forced crashes to see if they show up correctly.

* Initial implementation of CloudCityLogger

* Modify OkHttpInterceptor to use CloudCityLogger

And implement our own CloudCity OkHttp-compliant Logger to use in the
HttpLoggingInterceptor

EDIT: changed 'extends' to 'implements' because it's an interface and not a class

* Change logging in CloudCityOMNTApplication

And update TAG to be written out rather than using the getClass().getSimpleName()
since that will break the moment we start obfuscating our code.

* Replaced all occurances of Log.d() with CloudCityLogger.d() in cloudcity package

And changed all TAGs to be written out rather than using getClass().getSimpleName() since
that will break on the first release build.

* Replace the rest of the logs in the cloudcity package

Replaced the Log.v(), Log.i(), Log.e(), Log.w(), Log.wtf() with their
CloudCityLogger counterparts, but there's a small problem :)

All logs are now doubled, meaning that Timber is directly logging to logcat as well.
We should get rid of Log calls entirely.

* Get rid of android.util.Log calls from CloudCityLogger

To get rid of double-logging, we just need to call into Timber since
it directly logs to logcat as well. Maybe it could also write into a file,
but that's an improvement for later on.

* Update the OMNT codebase to use CloudCityLogger as well

* Get rid of Senty<->Timber integration

While the breadcrumbs are alright, each new log line was opening a new issue on Sentry
which isn't really doing us any favors if we're going to be debugging and diagnosing via logs.

* Remove android Log from CloudCityLogger