Address a TODO about NPE in LoggingServiceExtensions
And it makes sense because it's returning null,
i'm testing on a SIM-less phone, duuh

Also mark CellUtil::findRegisteredCell() as nullable because it can return null.

UPDATE: nope, restarting the phone makes it return a valid cell regardless of SIM.
this has nothing to do with the SIM at all, something else got messed up and died
in there, and we have no clue what other than the fact that cells (and other stuff)
were constantly null, regardless or force stopping and restarting the app.

Still, the newly added annotation is alright, because there *is* a possibility that
the method can return null.