Fix weird issue (or just me being SIM-less)
Somehow, the cell was always null when it was supposed to be remapped into the number.
This... shouldn't be because I don't have a SIM card. It used to work reliably regardless of that.
It still works reliably after restarting the phone.

But, lets add this constant to allow null cells for some random unknown special (test) cases.