Safety patch for Metric
Since streaming and reducing a list actually returns an Optional
which is either there or not, when trying to get() it or do something on it,
it throws and crashes the app if the Metric itself is empty.

Since there is no way to check for that from outside the Metric class, might
as well improve it's safety a bit.

This commit is probably unnecessary now, since it was a result of doing something
else wrong, which resulted in certain PingMonitor callbacks being called at inappropriate
times, and then it tried performing things on empty Metrics which always threw.

I figure this is a very decent fix/improvement regardless if the problem doesn't happen anymore.