Bit more updates to PingMonitor, it's nearing completion
Seems to work fine for now, just need to plug it in between the GPSMonitor and Iperf3Monitor

Also, needs a bit more cleaning up since there is some unused garbage in it that I thought
we'd need. However these ping tests work rather differently from the iPerf3 ones, which is
what the component was kinda based on.