Plug PingMonitor into GPSMonitor to start Iperf3 tests on successful pings
This thing works "fine" for now, but I'm not too fond of this approach.
The PingMonitor doesn't do time-based throttling, and will essentially ping
every 5 seconds (the GPSMonitor's value threshold) before succeding/failing
and the Iperf3 will just ignore it most of the time due to time-based throttling.

This is why I want to move it *into* the Iperf3 monitor's startDefault15sec test method itself.

So that the Iperf3Monitor is the single source of truth and does all the throttling,
and this guy just starting the ping test before the actual iperf3 test