Move PingMonitor out of the GPSMonitor and into the Iperf3Monitor
As the previous commit mentioned, the Iperf3Monitor is now the single source
of truth for running the iperf3, and throttling both on distance as well as
time constraints.

It will just enqueue the Iperf3Workers after a successful ping test, and if the
ping test fails, it will delay running the iperf3 test until the GPSMonitor pings him again;

that is, failed ping tests (and subsequent skipping of the iperf3 test run) should not count
towards time-based throttling.