Bugfix in the iperf3 monitor regarding whether test is running or not
Due to a slight oversight, the test was being set as running *before* the pingtest
begins; this breaks the whole "failed ping tests don't initiate iperf3 tests" idea,
since the test was being set as running while there was no actual iperf3 test being ran.

Addressed that situation so that the flag is set immediatelly before starting the test,
and not before the pingtest itself.

This fixes a case where the test was set as running, the pingtest failing, and the test
never really finishing because the iperf3 test never reaches a terminal state to unset the flag
due to it never really starting.