Fix bullshit problem in PingMonitor
Namely, we were just instantiating the pingParser to add a propertyChangeListener to it,
but since we couldn't get anything onto it (because that will be initialized later on in
the PingWorker instead), we were initializing it with a null BufferedReader.

Turns out there was another problem in the PingParser as well, but since the PingParser was changed
to 'allow' setting an 'external' PropertyChangeListener on it before it gets instantiated
- and then having it normally set it on itself upon next creation - the problem seems to have been avoided
and there are far less NPEs happening and PingWorker failures going on than before. Or perhaps that's just
a false-positive.

Either way, I consider this an improvement.

EDIT: added PingParser to this commit as well.