Handle retardation in Metric class
It's so tightly-coupled with it's UI, that it can't even update() itself
without touching the UI, which by the way - was not automatized, and does
not have to be there, so the whole update() won't work unless you followed
the "i guess you just have to know about it" recipe.

So, when a NPE is encountered in update() method on the UI part, it'll instantiate
the UI and re-try updating().