Change the gradually-increasing cancellation to just one that kills everything
The gradually increasing one worked fine, but was very much imperfect and, well, broken.
It worked fine in a sense that the system keeps on working and working and working whenever
it got stuck or locked up and managed to recover perfectly. At least the Iperf3Monitor did
and kept on working fine and reporting data.

It sucked and was broken in a sense that, ever second iperf3 test was *still running* when it
really wasn't. They were never cancelled from the workmanager, and were stuck in their whatever
stuck state they normally and usually got in. So if we nuke everything when the moment comes,
perhaps there won't be any lingering iperf3 tests broken somewhere in the background.

Who knows, maybe it's these fuckers that bring the library down?

EDIT: it's worth saying that this is being tested in extreme case of 0-second pause between the tests,
when things are *most* likely to get broken and messed up. So the failsafe is being tested
in such a harsh environment, not the "300-secs between tests" because aintnobodygottimeforthat.gif

EDIT2: added some logs to mention what's being done at which moment so we can analyze the situation better.