Fix iPerf3 API payload by moving the (signal/cell) measurements out of 'values'
The previous payload was 'bad' because the following things were being sent under iperf3 values:

```
                "cell_type": 4,
                "rsrp": -107,
                "rsrq": -17,
                "rssnr": -1
```

Actually, all "signal info" was being sent depending on what cell we're registered on, but that's besides the point.
This was due to the initial misunderstanding of what "cell info" actually is, but after implementing
that part properly i just didn't remove these values from the iperf3 values.

Now, they are no longer part of the iperf3 values (as they're also being sent as part of another API payload)
*but* we have moved them to the cell_info portion of the payload, so they can signify under what conditions
the iperf3 test was executed under.

Thus, the new iperf3 API payload will contain the signal info (previously contained in 'values') in the 'cell_info'
portion of the JSON, and keep the 'values' with just iperf3 measured values.

Something like this:
```
{
    "sensor_events": [
        {
            "accuracy": 13.89799976348877,
            "category": "Iperf3",
            "cell_info": {
                "cellId": 440,
                "cell_type": 4,
                "pci": 280,
                "rsrp": -106,
                "rsrq": -19,
                "rssnr": -1
            },
            "speed": 0,
            "lat": 44.7483276,
            "lon": 20.4258972,
            "values": {
                "download_last": 5.24156970255062,
                "download_max": 5.24156970255062,
                "download_mean": 3.8498429249998973,
                "download_median": 4.193507151416855,
                "download_min": 2.097791820143203,
                "ping_package_loss_last": 0,
                "ping_package_loss_max": 0,
                "ping_package_loss_mean": 0,
                "ping_package_loss_median": 0,
                "ping_package_loss_min": 0,
                "ping_RTT_last": 63.1,
                "ping_RTT_max": 45.96,
                "ping_RTT_mean": 63.1,
                "ping_RTT_median": 42.4,
                "ping_RTT_min": 38,
                "upload_last": 15.73576822905426,
                "upload_max": 15.73576822905426,
                "upload_mean": 8.800838381352749,
                "upload_median": 9.43533486239412,
                "upload_min": 0
            }
        }
    ]
}
```