Iperf3 "running status" Repository
Since I suspected that something is going very wrong with synchronization,
I made a centralized "is iperf3 test running" repository to be easily queried
before the Iperf3Monitor.startDefault15SecTest() is even called.

I did that because I notice *a lot* of tests "running" in the iperf3->instances
screen, as if a new test was started **exactly** every 5 seconds, which should
be impossible given the current synchronization locks in place. But yet, the
screen was completely misleading me there as well.

It turns out, that the real culprit of why iperf3 isn't running on PROD is because
pinging isn't enabled for it, and that right after the pingtest starts - and never
finishes successfully to let the listener know - it plants in fake data into the
middle-man databases used to convey data between the WM Workers running the native libs
and the rest of the app.

So, iperf3 test never really ever starts, pingtest never really ever finishes successfuly,
fake planted data tells us that the test is "still running", all of the code still remains
somewhat "locked in", waiting for the test to finish, and planted data keeps misleading us
that iperf3 tests, which are still ongoing, have been started **exactly** every 5 seconds.

This repository and PR shouldn't be needed, but helped pinpoint the exact problem more easily.